add common symmetric roll passes base class
