+++ pyroll/core/roll_pass/__init__.py
@@ -1,9 +1,11 @@
 from .base import BaseRollPass
-from .roll_pass import RollPass
+from .two_roll_pass import TwoRollPass
 from .three_roll_pass import ThreeRollPass
 from .deformation_unit import DeformationUnit
 
 from . import hookimpls
+
+RollPass = TwoRollPass
