+++ pyroll/core/roll_pass/hookimpls/__init__.py
@@ -1,4 +1,7 @@
-from . import roll_pass
+from . import base_roll_pass
+from . import symmetric_roll_pass
+from . import two_roll_pass
+from . import three_roll_pass
 from . import profile
 from . import roll
 from . import disk_element
